pyproject.toml
[project]
name = "musclemetric-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "werkzeug>=3.1.8",
]
